revise design of AstJson/AstToJsonConverter and align AstToJavascriptConverter with it [BC]
